+++ parachain/pallets/eth-app/src/mock.rs
@@ -148,7 +148,7 @@ where
 }
 
 parameter_types! {
-	pub const MaxMessagePayloadSize: u64 = 256;
+	pub const MaxMessagePayloadSize: u32 = 256;
 	pub const MaxMessagesPerCommit: u32 = 3;
 }
