+++ parachain/pallets/incentivized-channel/src/outbound/test.rs
@@ -116,7 +116,7 @@ impl pallet_assets::Config for Test {
 }
 
 parameter_types! {
-	pub const MaxMessagePayloadSize: u64 = 128;
+	pub const MaxMessagePayloadSize: u32 = 128;
 	pub const MaxMessagesPerCommit: u32 = 5;
 }
